Segregated backend functions to different files
